Fix assigning 0 to events on 64 bit system

+++ EventSystem.h
@@ -80,7 +80,7 @@ public:
 		callback = VoidCallbackWrapperCb;
 		return cb;
 	}
-	int operator =( int null )
+	size_t operator =( size_t null )
 	{
 		callback = (EventCallback)null;
 		return null;
